requirements.txt
streamlit
SpeechRecognition
pyttsx3
openai
python-dotenv
gtts
playsound